gemini-flashcards/requirements.txt
streamlit
google-generativeai
PyMuPDF
python-docx
pandas
